[RAPPS] separate application-view from gui.cpp
