+++ base/applications/rapps/include/misc.h
@@ -48,3 +48,5 @@ public:
 BOOL PathAppendNoDirEscapeW(LPWSTR pszPath, LPCWSTR pszMore);
 
 BOOL IsSystem64Bit();
+
+INT GetSystemColorDepth();
